+++ packages/dso-toolkit/src/components/tabs/tabs.models.ts
@@ -1,5 +1,6 @@
-export interface Tabs<TemplateFnReturnType> {
-  items: TabsItem<TemplateFnReturnType>[];
+export interface Tabs<TemplateFnRetunType> {
+  items: TabsItem<TemplateFnRetunType>[];
+  dsoTabSwitch?: (e: CustomEvent<TabsSwitchEvent>) => void;
 }
 
 export interface TabsItem<TemplateFnReturnType> {
@@ -7,4 +8,9 @@ export interface TabsItem<TemplateFnReturnType> {
   id: string;
   modifiers?: "active" | "disabled";
   content: TemplateFnReturnType | string;
+}
+
+export interface TabsSwitchEvent {
+  originalEvent: MouseEvent | KeyboardEvent;
+  selected: string;
 }